Add files via upload
This project takes in an input file where the first line indicates the number of test cases. Each test case is two lines in the input file where the first line indicates the amount of numbers to sum the squares and the second line is the numbers to sum. Note that the requirement for this project was to purposefully ignore negative numbers altogether. Another requirement for this was to use no loops or lists to complete this challenge.